requirements.txt
numpy
pandas
librosa
imageio
Pillow
matplotlib
